backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-MySQLdb==2.0.0
python-dotenv==1.0.0
PyJWT==2.11.0
Werkzeug==3.0.1
Pillow==10.1.0
python-magic==0.4.27
email-validator==2.1.0
reportlab==4.2.5
xhtml2pdf==0.2.16